skin-type-classifier/requirements.txt
torch
torchvision
Pillow
gradio==4.16.0
